requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
av
numpy
google-generativeai
pillow